refactoring to better support custom file share paths
