+++ fileglancer_central/model.py
@@ -1,0 +1,110 @@
+from datetime import datetime
+from typing import List, Optional, Dict
+
+from pydantic import BaseModel, Field, HttpUrl
+
+
+class FileSharePath(BaseModel):
+    """A file share path from the database"""
+    name: str = Field(
+        description="The name of the file share, which uniquely identifies the file share."
+    )
+    zone: str = Field(
+        description="The zone of the file share, for grouping paths in the UI."
+    )
+    group: Optional[str] = Field(
+        description="The group that owns the file share",
+        default=None
+    )
+    storage: Optional[str] = Field(
+        description="The storage type of the file share (home, primary, scratch, etc.)",
+        default=None
+    )
+    mount_path: str = Field(
+        description="The path where the file share is mounted on the local machine"
+    )
+    mac_path: Optional[str] = Field(
+        description="The path used to mount the file share on Mac (e.g. smb://server/share)",
+        default=None
+    )
+    windows_path: Optional[str] = Field(
+        description="The path used to mount the file share on Windows (e.g. \\\\server\\share)",
+        default=None
+    )
+    linux_path: Optional[str] = Field(
+        description="The path used to mount the file share on Linux (e.g. /unix/style/path)",
+        default=None
+    )
+
+class FileSharePathResponse(BaseModel):
+    paths: List[FileSharePath] = Field(
+        description="A list of file share paths"
+    )
+    
+class TicketComment(BaseModel):
+    """A comment on a ticket"""
+    author_name: str = Field(
+        description="The author of the comment"
+    )
+    author_display_name: str = Field(
+        description="The display name of the author"
+    )
+    body: str = Field(
+        description="The body of the comment"
+    )
+    
+class TicketComment(BaseModel):
+    """A comment on a ticket"""
+    author_name: str = Field(
+        description="The author of the comment"
+    )
+    author_display_name: str = Field(
+        description="The display name of the author"
+    )
+    body: str = Field(
+        description="The body of the comment"
+    )
+    created: datetime = Field(
+        description="The date and time the comment was created"
+    )
+    updated: datetime = Field(
+        description="The date and time the comment was updated"
+    )
+
+class Ticket(BaseModel):
+    """A JIRA ticket"""
+    key: str = Field(
+        description="The key of the ticket"
+    )
+    created: datetime = Field(
+        description="The date and time the ticket was created"
+    )
+    updated: datetime = Field(
+        description="The date and time the ticket was updated"
+    )
+    status: str = Field(
+        description="The status of the ticket"
+    )
+    resolution: str = Field(
+        description="The resolution of the ticket"
+    )
+    description: str = Field(
+        description="The description of the ticket"
+    )
+    link: HttpUrl = Field(
+        description="The link to the ticket"
+    )
+    comments: List[TicketComment] = Field(
+        description="The comments on the ticket"
+    )
+    
+
+class UserPreference(BaseModel):
+    """A user preference"""
+    key: str = Field(
+        description="The key of the preference"
+    )
+    value: Dict = Field(
+        description="The value of the preference"
+    )
+
